+++ frontends/web/ts/gb.ts
@@ -614,7 +614,7 @@ export class GameboyEmulator extends EmulatorBase implements Emulator {
             ch1: output[1],
             ch2: output[2],
             ch3: output[3],
-            ch4: 0
+            ch4: output[4]
         };
     }