+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for audio channel 4 (noise) 🔈
+* Better trigger support for audio channels 🔈
 
 ### Changed
 
@@ -17,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-* Envelope support for both channel 2 and 4
+* Envelope support for both channel 2 and 4 🔈
 
 ## [0.7.3] - 2023-04-02