+++ CHANGELOG.md
@@ -14,11 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-*
+* Added CH4 public API method for WASM
 
 ### Fixed
 
 * Envelope support for both channel 2 and 4 🔈
+* Issue related to the wave length stop flag 🔈
 
 ## [0.7.3] - 2023-04-02